chore: Auslagerung von FileReader in eigene Komponente
